+++ R/aio.R
@@ -234,10 +234,10 @@ call_aio <- function(aio) invisible(.Call(rnng_aio_call, aio))
 #'
 call_aio_ <- function(aio) invisible(.Call(rnng_wait_thread_create, aio))
 
-#' Collect the Value of an Aio or List of Aios
+#' Retrieve Data of an Aio or List of Aios
 #'
-#' \code{collect_aio} retrieves the value of a single, or list of, asynchronous
-#'     Aio operations, waiting for their completion if still in progress.
+#' \code{aio_data} retrieves the data value of an Aio, or list of Aios, waiting
+#'     for resolution if still in progress.
 #'
 #' @param x an Aio or list of Aios (objects of class \sQuote{sendAio},
 #'     \sQuote{recvAio} or \sQuote{ncurlAio}).
@@ -245,19 +245,22 @@ call_aio_ <- function(aio) invisible(.Call(rnng_wait_thread_create, aio))
 #' @return Depending on the type of \sQuote{x} supplied, either an object or a
 #'     list of objects (the same length as \sQuote{x}, preserving names).
 #'
+#' @details This function will wait for the asynchronous operation(s) to
+#'     complete if still in progress (blocking).
+#'
 #' @export
 #'
-collect_aio <- function(x) .Call(rnng_aio_collect, x)
+aio_data <- function(x) .Call(rnng_aio_data, x)
 
-#' Collect the Value of a List of Aios
+#' Retrieve Data of an Aio or List of Aios
 #'
-#' \code{collect_aio_} is a variant that allows user interrupts, suitable for
+#' \code{aio_data_} is a variant that allows user interrupts, suitable for
 #'     interactive use.
 #'
-#' @rdname collect_aio
+#' @rdname aio_data
 #' @export
 #'
-collect_aio_ <- function(x) .Call(rnng_aio_collect_safe, x)
+aio_data_ <- function(x) .Call(rnng_aio_data_safe, x)
 
 #' Stop Asynchronous Aio Operation
 #'